# backend/cmd/api/docs/swagger.yaml
definitions:
  types.CampaignResponse:
    properties:
      district:
        type: string
      id:
        type: integer
      year:
        type: integer
    type: object
  types.CreatePromiseUpdateRequest:
    properties:
      note:
        type: string
      updated_at:
        type: string
    type: object
  types.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  types.LoginResponse:
    properties:
      role:
        type: string
      token:
        type: string
    type: object
  types.PoliticianResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      party:
        type: string
      politician_code:
        type: string
    type: object
  types.PromiseDetailResponse:
    properties:
      announced_at:
        type: string
      campaign:
        $ref: '#/definitions/types.CampaignResponse'
      detail:
        type: string
      id:
        type: integer
      politician:
        $ref: '#/definitions/types.PoliticianResponse'
      status:
        type: string
      updates:
        items:
          $ref: '#/definitions/types.PromiseUpdateResponse'
        type: array
    type: object
  types.PromiseResponse:
    properties:
      announced_at:
        type: string
      campaign:
        $ref: '#/definitions/types.CampaignResponse'
      detail:
        type: string
      id:
        type: integer
      politician:
        $ref: '#/definitions/types.PoliticianResponse'
      status:
        type: string
    type: object
  types.PromiseUpdateResponse:
    properties:
      id:
        type: integer
      note:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: This is the API documentation for  project.
  termsOfService: http://swagger.io/terms/
  title: API documentation
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Sets up the routes for authentication
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Auth Routes
      tags:
      - Auth
  /campaigns/:
    get:
      consumes:
      - application/json
      description: get list of campaigns
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.CampaignResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List Campaigns
      tags:
      - campaigns
  /cats/:
    get:
      consumes:
      - application/json
      description: get list of cats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: Get Cats
      tags:
      - cats
  /politicians/:
    get:
      consumes:
      - application/json
      description: get list of politicians
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.PoliticianResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List Politicians
      tags:
      - politicians
  /politicians/{id}/promises:
    get:
      consumes:
      - application/json
      description: get list of promises by politician ID
      parameters:
      - description: Politician ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.PromiseResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Promises By Politician
      tags:
      - politicians
  /promises/{id}:
    get:
      consumes:
      - application/json
      description: get promise detail by ID
      parameters:
      - description: Promise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PromiseDetailResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Promise Detail
      tags:
      - promises
  /promises/{id}/updates:
    post:
      consumes:
      - application/json
      description: add an update to a promise (admin only)
      parameters:
      - description: Promise ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create Promise Update Request
        in: body
        name: createPromiseUpdateRequest
        required: true
        schema:
          $ref: '#/definitions/types.CreatePromiseUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add Promise Update
      tags:
      - promises
  /promises/all:
    get:
      consumes:
      - application/json
      description: get list of all promises
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.PromiseResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List All Promises
      tags:
      - promises
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"